Ran some package updates
I ran some updates to try and fix some security issues however its a pretty deep rooted issue mostly involving the packages themselves needing updates from the developers.

I spent a lot of time on this and likely broke something so it got its own branch. Try pulling this branch and playing with it, but I'm not sure if you update packages on one branch if it will affect other branches as well