Refactor Ability Repository for proper expiry
When we set a local storage for expiry, it creates a new record, while this is nice for big blobs (all pokemon), we'd need to create one for each individual object (specific pokemon, ability, move etc). I feel its better to store this in the record itself rather than create a new UNIQUE entry because
a. it would require a unique name for each (Pokemon and Ability clash)
b. since a, looking at a new ability would NOT create a new expiry for it
c. when checking for expiry, it wouldn't be able to find it, thus the repository was failing the lookup and fetching the ability all over again
What I did was rewrote the repository to STORE the expiry with the object, and will update whenever an update is made to it (use case? not sure yet, maybe if the app/api makes an update that we need to run in the background we can update the repository). A further update to PokemonRepository would need to be made because individual pokemon lookup is not using it right now, only the blob at the start of the application run is